// src/app/models/player.model.ts
export class PlayerModel {
  account_plan: string;
  account_verified: any;
  birthday: string;
  country: string;
  edad: any;
  email: string;
  firstname: string;
  genre: string;
  id: any;
  imagen: string;
  lastname: string;
  phone: string;
  picture: string;
  player_club: string;
  player_division: string;
  player_foot: string;
  player_height: string;
  player_number: string;
  player_position: string;
  province: string;
  sexo: string;
  state: string;
  stripe_customer: string;
}